# Added map_samples implementation