--- a/fiftyone/core/dataset.py
+++ b/fiftyone/core/dataset.py
@@ -15,9 +15,9 @@ import itertools
 import logging
 import numbers
 import os
+from queue import Queue
 import random
 import string
-import multiprocessing as mp
 
 
 from bson import json_util, ObjectId, DBRef
@@ -2895,6 +2895,8 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
         autosave=False,
         batch_size=None,
         batching_strategy=None,
+        skip_failures=True,
+        warn_failures=False,
     ):
         """Maps a function over the samples in the dataset.
 
@@ -2910,6 +2912,10 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
             autosave (False): whether to automatically save the results
             batch_size (None): the batch size to use when autosaving samples
             batching_strategy (None): the batching strategy to use for each save
+            skip_failures (True): whether to gracefully continue without raising an
+                error if processing fails for a sample
+            warn_failures (False): whether to log a warning if processing fails for
+                a sample
 
         Returns:
             the result of the mapping operation, which is a list of the results
@@ -2919,10 +2925,8 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
             raise ValueError("map_func must be callable")
 
         # Create a thread pool
-        num_workers = min(num_workers, mp.cpu_count() - 1)
         with ThreadPoolExecutor(max_workers=num_workers) as executor:
-            futures = []
-
+            futures_queue = Queue()
             # Submit samples to the worker pool
             for sample in self.iter_samples(
                 progress=progress,
@@ -2931,9 +2935,24 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
                 batching_strategy=batching_strategy,
             ):
                 future = executor.submit(map_func, sample, map_func_args)
-                futures.append(future)
+                futures_queue.put(future)
 
-            result = [future.result() for future in futures]
+            # Process results
+            result = []
+            for _ in range(futures_queue.qsize()):
+                try:
+                    future = futures_queue.get()
+                    result.append(future.result())
+                except Exception as e:
+                    if not skip_failures:
+                        raise RuntimeError(
+                            "Worker failed while processing sample"
+                        ) from e
+
+                    if warn_failures:
+                        logger.warning("Error processing sample: %s", e)
+
+                    result.append(None)
 
             if aggregate_func is not None:
                 if callable(aggregate_func):