--- a/fiftyone/core/dataset.py
+++ b/fiftyone/core/dataset.py
@@ -15,7 +15,6 @@ import itertools
 import logging
 import numbers
 import os
-from queue import Queue
 import random
 import string
 
@@ -2926,7 +2925,10 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
 
         # Create a thread pool
         with ThreadPoolExecutor(max_workers=num_workers) as executor:
-            futures_queue = Queue()
+            futures = []
+            result = []
+            max_queue = 10000
+
             # Submit samples to the worker pool
             for sample in self.iter_samples(
                 progress=progress,
@@ -2935,24 +2937,17 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
                 batching_strategy=batching_strategy,
             ):
                 future = executor.submit(map_func, sample, map_func_args)
-                futures_queue.put(future)
+                futures.append(future)
+                if len(futures) > max_queue:
+                    self._process_future_result(
+                        futures, result, skip_failures, warn_failures
+                    )
+                    futures = []
 
-            # Process results
-            result = []
-            for _ in range(futures_queue.qsize()):
-                try:
-                    future = futures_queue.get()
-                    result.append(future.result())
-                except Exception as e:
-                    if not skip_failures:
-                        raise RuntimeError(
-                            "Worker failed while processing sample"
-                        ) from e
-
-                    if warn_failures:
-                        logger.warning("Error processing sample: %s", e)
-
-                    result.append(None)
+            # Process remaining results
+            self._process_future_result(
+                futures, result, skip_failures, warn_failures
+            )
 
             if aggregate_func is not None:
                 if callable(aggregate_func):
@@ -2961,6 +2956,23 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
                     raise ValueError("aggregate_func must be callable")
 
             return result
+
+    def _process_future_result(
+        self, futures, result, skip_failures, warn_failures
+    ):
+        for future in futures:
+            try:
+                result.append(future.result())
+            except Exception as e:
+                if not skip_failures:
+                    raise RuntimeError(
+                        "Worker failed while processing sample"
+                    ) from e
+
+                if warn_failures:
+                    logger.warning("Error processing sample: %s", e)
+
+                result.append(None)
 
     def iter_samples(
         self,