## What changes are proposed in this pull request?

* Added `map_samples` for parallel processing in conjunction with iter samples
* Sample implementation of `map_samples` using `evaluate_detections` 

## How is this patch tested? If it is not, please explain why.

* Tested local SDK against remote MongoDB 
* Script to run experiment: https://gist.github.com/minhtuev/834e2efb1c1538097912b7aa07ee5a63

Example benchmark code with evaluate_detections 

```
import time
import fiftyone as fo

def evaluate_with_workers(dataset, num_workers_list):
    """
    Evaluates detections for different numbers of workers and measures execution time.
    
    Args:
        dataset: FiftyOne dataset to evaluate
        num_workers_list: List of worker counts to test
    """
    for num_workers in num_workers_list:
        print(f"\nEvaluating detections with num_workers={num_workers}")
        
        start_time = time.time()
        try:
            results = dataset.evaluate_detections(
                "detections",
                gt_field="detections",
                eval_key=f"eval_key_{num_workers}",
                num_workers=num_workers
            )
            elapsed = time.time() - start_time
            print(f"Elapsed time: {elapsed:.2f} seconds")
            
        except Exception as e:
            print(f"Error occurred with {num_workers} workers: {str(e)}")

def main():
    # Load dataset
    dataset = fo.load_dataset("bdd100k-validation")
    
    # Test with different worker counts
    worker_counts = [None, 2, 4, 8, 16]
    evaluate_with_workers(dataset, worker_counts)

if __name__ == "__main__":
    main()
```

Sample result:
```
Evaluating detections with num_workers=None
Evaluating detections...
 100% |████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 8000/8000 [12.7m elapsed, 0s remaining, 10.3 samples/s]
Elapsed: 770.9798429012299
Evaluating detections with num_workers=2
Evaluating detections...
 100% |████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 8000/8000 [2.7m elapsed, 0s remaining, 73.2 samples/s]
Elapsed: 178.74396586418152
Evaluating detections with num_workers=4
Evaluating detections...
 100% |████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 8000/8000 [2.9m elapsed, 0s remaining, 68.2 samples/s]
Elapsed: 200.93079495429993
Evaluating detections with num_workers=8
Evaluating detections...
 100% |████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 8000/8000 [3.2m elapsed, 0s remaining, 54.2 samples/s]
Elapsed: 202.46744012832642
Evaluating detections with num_workers=16
Evaluating detections...
 100% |████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 8000/8000 [4.4m elapsed, 0s remaining, 70.4 samples/s]
Elapsed: 273.4626371860504
```

## Release Notes

### Is this a user-facing change that should be mentioned in the release notes?

<!--
Please fill in relevant options below with an "x", or by clicking the checkboxes
after submitting this pull request. Example:
-   [x] Selected option
-->

-   [ ] No. You can skip the rest of this section.
-   [ ] Yes. Give a description of this change to be included in the release
        notes for FiftyOne users.

(Details in 1-2 sentences. You can just refer to another PR with a description
if this PR is part of a larger change.)

### What areas of FiftyOne does this PR affect?

-   [ ] App: FiftyOne application changes
-   [ ] Build: Build and test infrastructure changes
-   [x] Core: Core `fiftyone` Python library changes
-   [ ] Documentation: FiftyOne documentation changes
-   [ ] Other


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **New Features**
	- Added `map_samples` method to `Dataset` and `DatasetView` classes, enabling concurrent processing of samples.
	- Enhanced detection evaluation with parallel processing and improved sample processing logic.
	- Introduced new utility functions for more efficient sample evaluation.

- **Performance Improvements**
	- Implemented thread pool-based sample mapping.
	- Added support for multi-threaded sample processing in detection evaluations.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->