@swheaton 2025-01-28
@brimoor disagree, it is going in the right direction. Namely it's a simple first step and has 4x improvement on normal `iter_samples()` for this one test (as mentioned above need more testing before merging).

Bringing everything into memory is a problem I agree so we'll have to fix that.

But everything else you describe are ideas for additional improvement which we can explore next.

"Perfect is the enemy of good" - I don't think all these things need to be in place before we merge something.