@minhtuev 2025-01-28
@brimoor : yes agreed with @swheaton , we started with Thread parallelism because:

1. It is easier to implement and we do not have to deal with the issue of pickling/unpickling FO objects. 
2. We can review if this is the interface that we want.

I want to run a couple more experiments with `evaluate_detections` and other `evaluate_*` methods to show that this is the desired paradigm that we can adopt across the board to give a performance lift. 

We can also add an option for process-parallelism underneath `map_samples` so that users can choose between thread vs. process; process parallelism is totally feasible if we do not need to pass FO objects around.