@brimoor 2025-01-29
> @brimoor disagree, it is going in the right direction. Namely it's a simple first step and has 4x improvement on normal `iter_samples()` for this one test (as mentioned above need more testing before merging).
> 
> Bringing everything into memory is a problem I agree so we'll have to fix that.
> 
> But everything else you describe are ideas for additional improvement which we can explore next.
> 
> "Perfect is the enemy of good" - I don't think all these things need to be in place before we merge something.

`iter_samples(autosave=True)` is an ideal candidate for distributed computing. Each worker can read + compute + write its own data in isolation. There is no need to read sample objects in the main thread and pickle them. 

The current pattern in this PR won't support swapping out multithreading or multiprocessing for a Hadoop cluster, for example. In a distributed setting, every worker *definitely* needs to do its own I/O.