@kaixi-wang 2025-01-29
For the parallelization I think Brian is describing, we would need to initialize the mongo client from within map_samples (this would also allow us to make more connections to mongo since mongo limits the total number of connections per client). Then call iter_samples inside the worker function on a view defined by the args. 

To avoid serializing a long list of ids from a values() call (which itself can be expensive I/O), but not rely on the order of samples returned in slices defined by skip/limit (which may not be stable),  I think we can try defining the view from which we iterate on by passing a min/max _id so each worker starts with a match stage on an indexed field, which will also help with performance. 

There will be some overhead in getting the bounds upfront, but it's hopefully negligible