@minhtuev 2025-01-29
Updated implementation of thread execution using dedicated threads with input and output queue to minimize resource allocations/deallocations

```
Results Summary:
----------------------------------------------------------------------
   Workers | Batch Size |   Time (s) |  Completed
----------------------------------------------------------------------
        12 |          8 |     122.80 |        Yes
         2 |          4 |     146.66 |        Yes
        16 |          8 |     162.61 |        Yes
         8 |          8 |     188.54 |        Yes
        12 |          2 |     198.37 |        Yes
         1 |          4 |     199.56 |        Yes
         8 |          1 |     200.04 |        Yes
         4 |          1 |     201.11 |        Yes
        16 |          1 |     201.25 |        Yes
         4 |          2 |     202.36 |        Yes
        12 |          1 |     203.99 |        Yes
         2 |          2 |     204.63 |        Yes
         8 |          2 |     204.77 |        Yes
         1 |          2 |     206.53 |        Yes
         4 |       None |     208.45 |        Yes
        12 |       None |     208.63 |        Yes
         1 |       None |     209.12 |        Yes
         4 |          8 |     212.51 |        Yes
         1 |          8 |     213.84 |        Yes
        16 |          2 |     215.00 |        Yes
         2 |          8 |     215.75 |        Yes
        16 |       None |     215.94 |        Yes
         8 |       None |     216.11 |        Yes
         4 |          4 |     216.12 |        Yes
         2 |       None |     221.50 |        Yes
         2 |          1 |     223.68 |        Yes
         8 |          4 |     226.68 |        Yes
        12 |          4 |     229.27 |        Yes
        16 |          4 |     231.98 |        Yes
         1 |          1 |     241.54 |        Yes
----------------------------------------------------------------------

Experiment Note: using dedicated threads to minimize resources allocation
```