@swheaton 2025-01-29
> For the parallelization I think Brian is describing, we would need to initialize the mongo client from within map_samples (this would also allow us to make more connections to mongo since mongo limits the total number of connections per client). Then call iter_samples inside the worker function on a view defined by the args.
> 
> To avoid serializing a long list of ids from a values() call (which itself can be expensive I/O), but not rely on the order of samples returned in slices defined by skip/limit (which may not be stable), I think we can try defining the view from which we iterate on by passing a min/max _id so each worker starts with a match stage on an indexed field, which will also help with performance.
> 
> There will be some overhead in getting the bounds upfront, but it's hopefully negligible

Yeah this is what concerned me and why I thought we should read in a single thread and farm out the objects elsewhere. Until we prove we must go further in order to improve performance, then we will have to start making these tradeoffs.

You have a good idea though. But how do you split 10M ID's into N batches defined by [min_id, max_id]? We would have to do a full IXscan I think, UNLESS we make an assumption about ID distribution, which I don't think we can.