@kaixi-wang 2025-01-29
> But how do you split 10M ID's into N batches defined by [min_id, max_id]? We would have to do a full IXscan I think, UNLESS we make an assumption about ID distribution, which I don't think we can.

That might be a spike. But fwiw, an IX scan on _id for 10 million samples takes less than a second. 
I agree we definitely cannot make an assumption on distribution without additional metadata that doesn't exist now