@swheaton 2025-01-30
> > But how do you split 10M ID's into N batches defined by [min_id, max_id]? We would have to do a full IXscan I think, UNLESS we make an assumption about ID distribution, which I don't think we can.
> 
> That might be a spike. But fwiw, an IX scan on _id for 10 million samples takes less than a second. I agree we definitely cannot make an assumption on distribution without additional metadata that doesn't exist now

Still don't know why `hint` has such an impact but I'll take it. Is this still true if sorted by `_id`, I assume? This could be a good approach then, I like it. We wouldn't have to store all IDs just rip through and create batches of rough size.